added basic virtual machine engine.
we now have a run_file function that doesn't run a file,
but it does execute virtual bytecodes we have stored in
the program array.

it uses a variant of a direct threading method of flow control,
because we use a function table to map opcodes -> goto labels.
so I suppose it could be called indirect? regardless it should
work quite fast. there's a bug to add jmp opcodes, as well as
file loading (which will likely happen after more opcodes are
added.)